% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{accuracy}
\alias{accuracy}
\title{Compute the Bias and MSE of the Estimates}
\usage{
accuracy(data)
}
\arguments{
\item{data}{Named list containing the parameter names under \code{"parameters"},
the true parameter values under \code{"simulated"}, and the estimated 
parameter values under \code{"estimated"}. Typically the result of 
\code{\link[paramrel]{prepare}}}
}
\value{
Dataframe containing the parameter (\code{"parameter"}), the 
bias, and the MSE
}
\description{
This function depends on the assumption that the order of the estimated 
parameters is the same as for the generating parameters. This is automatically
true in the \code{\link[paramrel]{execute_study}} function.
}
