% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{arx-class}
\alias{arx-class}
\alias{arx}
\title{An S4 Class to Represent the ARX(1) Model.}
\description{
An S4 Class to Represent the ARX(1) Model.
}
\details{
Defines the \code{arx} class, which states that for a given dependent
variable \eqn{y}, the previous values of this variable in part determine 
current values of the variable:

\deqn{y_{t} = a + by_{t - 1} + cx_{t}}

This class is solely defined by its parameters \eqn{a}, \eqn{b} and \eqn{c}, 
which should all be provided through the construction of this class.
}
\section{Slots}{

\describe{
\item{\code{parameters}}{Numeric vector containing the values of the parameters of 
the model, namely \eqn{a}, \eqn{b}, and \eqn{c} in this order. If left 
unspecified, the model will default to \code{c(0, 0, 0)}.}
}}

