% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{autoregressive-class}
\alias{autoregressive-class}
\alias{autoregressive}
\title{An S4 Class to Represent the Autoregressive Model.}
\description{
An S4 Class to Represent the Autoregressive Model.
}
\details{
Defines the \code{autoregressive} class, which states that for a given dependent
variable \eqn{y}, the previous values of this variable in part determine 
current values of the variable:

\deqn{y_{t} = a + by_{t - 1}}

This class is solely defined by its parameters \eqn{a} and \eqn{b}, which 
should all be provided through the construction of this class.
}
\section{Slots}{

\describe{
\item{\code{parameters}}{Numeric vector containing the values of the parameters of 
the model, namely \eqn{a} and \eqn{b} in this order. If left unspecified, the 
model will default to \code{c(0, 0)}.}
}}

