% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{bounds}
\alias{bounds}
\alias{bounds,linear-method}
\alias{bounds,quadratic-method}
\alias{bounds,cubic-method}
\alias{bounds,main_effect-method}
\alias{bounds,interaction-method}
\alias{bounds,ar1-method}
\alias{bounds,arx-method}
\title{Get the Bounds of the Model}
\usage{
bounds(model, ...)

\S4method{bounds}{linear}(model, intercept = c(-10, 10), slope = c(-5, 5))

\S4method{bounds}{quadratic}(model, intercept = c(-10, 10), slope = c(-5, 5))

\S4method{bounds}{cubic}(model, intercept = c(-10, 10), slope = c(-5, 5))

\S4method{bounds}{main_effect}(model, intercept = c(-10, 10), slope = c(-5, 5))

\S4method{bounds}{interaction}(model, intercept = c(-10, 10), slope = c(-5, 5))

\S4method{bounds}{ar1}(model, intercept = c(-10, 10), slope = c(-5, 5))

\S4method{bounds}{arx}(model, intercept = c(-10, 10), slope = c(-5, 5))
}
\arguments{
\item{model}{Object of one of the different model classes (e.g., linear, 
quadratic,...)}

\item{...}{Additional arguments provided to the model of choice.}

\item{intercept}{Numeric vector with two values denoting the lower and upper
bound of the intercept parameter. Defaults to \code{c(-10, 10)}.}

\item{slope}{Numeric vector with two values denoting the lower and upper 
bound of all slope parameters for the covariates. Defaults to \code{c(-5, 5)}.}
}
\value{
Matrix where the left/right column specifies the lower/upper bounds 
of the parameters in the order specified in the description of the models
}
\description{
Get the Bounds of the Model
}
\examples{
# Example given for the linear model, but also possible for the other models
bounds(linear())

bounds(
  linear(),
  intercept = c(-10, 10),
  slope = c(0, 10)
)

}
