% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{cubic-class}
\alias{cubic-class}
\alias{cubic}
\title{An S4 Class to Represent the Cubic Model.}
\description{
An S4 Class to Represent the Cubic Model.
}
\details{
Defines the \code{cubic} class, which states that for a given input \eqn{x}, 
the relationship between \eqn{x} and the dependent variable \eqn{y} is 
quadratic:

\deqn{y = a + bx + cx^2 + dx^3}

This class is solely defined by its parameters \eqn{a}, \eqn{b}, \eqn{c}, and
\eqn{d}, which should be provided through the construction of this class.
}
\section{Slots}{

\describe{
\item{\code{parameters}}{Numeric vector containing the values of the parameters of 
the model, namely \eqn{a}, \eqn{b}, and \eqn{c} in this order. If left 
unspecified, the model will default to \code{c(0, 0, 0, 0)}.}

\item{\code{sd}}{Numeric defining the error around the deterministic part defined by 
the slot \code{parameters}. If left unspecified, will default to \code{1}}
}}

