% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{descriptives}
\alias{descriptives}
\title{Compute the Descriptive Statistics for the Parameters}
\usage{
descriptives(data)
}
\arguments{
\item{data}{Named list containing the parameter names under \code{"parameters"},
the true parameter values under \code{"simulated"}, and the estimated 
parameter values under \code{"estimated"}. Typically the result of 
\code{\link[paramrel]{prepare}}}
}
\value{
Dataframe containing the parameter and bin (\code{"parameter"}, 
\code{"bin"}), as well as the summary statistics for each.
}
\description{
Descriptive statistics include the mean, standard deviation, median, and 95%
confidence interval across participants. Statistics are computed per bin-type.
}
