% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\alias{estimate,linear-method}
\alias{estimate,quadratic-method}
\alias{estimate,main-method}
\alias{estimate,interaction-method}
\alias{estimate,ar1-method}
\alias{estimate,arx-method}
\title{Estimate a Model from Data}
\usage{
estimate(model, ...)

\S4method{estimate}{linear}(model, data)

\S4method{estimate}{quadratic}(model, data)

\S4method{estimate}{main}(model, data)

\S4method{estimate}{interaction}(model, data)

\S4method{estimate}{ar1}(model, data)

\S4method{estimate}{arx}(model, data)
}
\arguments{
\item{model}{Object representing the model to be used. Note that this may just
be the class name and does not need any particular parameters to be specified.}

\item{data}{Data.frame containing the relevant variables, specifically 
\code{"y"}, \code{"x"}, \code{"z"}, and \code{"time"}. Ideally, all variables
are present in the data.frame, even when they are not applicable to a 
particular model.}
}
\value{
List containing the result of \code{lm} under \code{"fit"} and the 
model with estimated parameters under \code{"model"}
}
\description{
Estimation of the different models from a particular data set. Uses the 
\code{lm} function under the hood.
}
