% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_retest.R
\name{generate_parameters}
\alias{generate_parameters}
\title{Generate a Random Parameter Set}
\usage{
generate_parameters(n, mean, sd)
}
\arguments{
\item{n}{Integer denoting the number of parameters to generate}

\item{mean}{Numeric vector containing the means of the parameters}

\item{sd}{Numeric denoting the (shared) standard deviation of the parameters}
}
\value{
Numeric matrix of size n x k, where n is the number of randomly
generated parameter sets and k the number of parameters of the model
}
\description{
Generate a Random Parameter Set
}
