% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_retest.R
\name{generate_testretest_study}
\alias{generate_testretest_study}
\title{Generate Complete Test-Retest Study Design}
\usage{
generate_testretest_study(study_size = "pilot")
}
\arguments{
\item{study_size}{Size of study ("pilot", "medium", "full")}
}
\value{
List of study conditions
}
\description{
Generate Complete Test-Retest Study Design
}
