% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{icc}
\alias{icc}
\title{Compute the \eqn{ICC(A, 1)} for Participant and Bin.}
\usage{
icc(data)
}
\arguments{
\item{data}{Named list containing the parameter names under \code{"parameters"},
the true parameter values under \code{"simulated"}, and the estimated 
parameter values under \code{"estimated"}. Typically the result of 
\code{\link[paramrel]{prepare}}}
}
\value{
Dataframe containing the parameter (\code{"parameter"}), systematic 
and unsystematic variances (\code{"systematic"}, \code{"unsystematic"}), and 
the estimated \eqn{ICC(A, 1)} for each parameter (\code{"icc"})
}
\description{
In our study, we define the \eqn{ICC(A, 1)} as:

\deqn{ICC(A, 1) = \frac{\sigma_\text{participants}}^2}{\sigma_\text{total}^2}

where \eqn{\sigma^2} denotes the variance. This \eqn{ICC} puts the systematic 
variance -- that is, the interindividual differences in the values of the 
parameters -- against the total observed variance in the parameters, checking 
whether individual differences can be reliably picked up on.
}
