% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{initialize,ar1-method}
\alias{initialize,ar1-method}
\title{Constructor for the \code{\link[paramrel]{ar1-class}}}
\usage{
\S4method{initialize}{ar1}(.Object, parameters = c(0, 0), sd = 1)
}
\arguments{
\item{parameters}{Numeric vector containing the values of the parameters of 
the model, namely \eqn{a} and \eqn{b} in this order. If left unspecified, the 
model will default to \code{c(0, 0)}.}

\item{sd}{Numeric defining the error around the deterministic part defined by 
the argument \code{parameters}. If left unspecified, will default to \code{1}}
}
\description{
Constructor for the \code{\link[paramrel]{ar1-class}}
}
