% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{initialize,main-method}
\alias{initialize,main-method}
\title{Constructor for the \code{\link[paramrel]{main-class}}}
\usage{
\S4method{initialize}{main}(.Object, parameters = c(0, 0, 0))
}
\arguments{
\item{parameters}{Numeric vector containing the values of the parameters of 
the model, namely \eqn{a}, \eqn{b}, and \eqn{c} in this order. If left 
unspecified, the model will default to \code{c(0, 0, 0)}.}
}
\description{
Constructor for the \code{\link[paramrel]{main-class}}
}
