% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{initialize,main_effect-method}
\alias{initialize,main_effect-method}
\title{Constructor for the \code{\link[paramrel]{main_effect-class}}}
\usage{
\S4method{initialize}{main_effect}(.Object, parameters = c(0, 0, 0), sd = 1)
}
\arguments{
\item{parameters}{Numeric vector containing the values of the parameters of 
the model, namely \eqn{a}, \eqn{b}, and \eqn{c} in this order. If left 
unspecified, the model will default to \code{c(0, 0, 0)}.}

\item{sd}{Numeric defining the error around the deterministic part defined by 
the argument \code{parameters}. If left unspecified, will default to \code{1}}
}
\description{
Constructor for the \code{\link[paramrel]{main_effect-class}}
}
