% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{initialize,model-method}
\alias{initialize,model-method}
\title{Constructor for the \code{\link[paramrel]{model-class}}}
\usage{
\S4method{initialize}{model}(.Object, parameters = numeric(0), sd = 1)
}
\arguments{
\item{.Object}{Model to initialize}

\item{parameters}{Numeric vector containing the values of the parameters of 
the model. If left unspecified, an empty numeric vector.}

\item{sd}{Numeric defining the error around the deterministic part defined by 
the argument \code{parameters}. If left unspecified, will default to \code{1}}
}
\description{
Constructor for the \code{\link[paramrel]{model-class}}
}
