% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{initialize,quadratic-method}
\alias{initialize,quadratic-method}
\title{Constructor for the \code{\link[paramrel]{quadratic-class}}}
\usage{
\S4method{initialize}{quadratic}(.Object, parameters = c(0, 0, 0), sd = 1)
}
\arguments{
\item{parameters}{Numeric vector containing the values of the parameters of 
the model, namely \eqn{a}, \eqn{b}, and \eqn{c} in this order. If left 
unspecified, the parameters will default to \code{c(0, 0, 0)}.}
}
\description{
Constructor for the \code{\link[paramrel]{quadratic-class}}
}
