% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{interaction-class}
\alias{interaction-class}
\alias{interaction}
\title{An S4 Class to Represent the Interaction Model.}
\description{
An S4 Class to Represent the Interaction Model.
}
\details{
Defines the \code{interaction} class, which states that for two given input
variables \eqn{x} and \eqn{z}, the relationship between these variables and 
the dependent variable \eqn{y} is one with a main and interaction effect:

\deqn{y = a + bx + cz + dxz}

This class is solely defined by its parameters \eqn{a}, \eqn{b}, \eqn{c},
and \eqn{d}, which should all be provided through the construction of this 
class.
}
\section{Slots}{

\describe{
\item{\code{parameters}}{Numeric vector containing the values of the parameters of 
the model, namely \eqn{a}, \eqn{b}, \eqn{c}, and \eqn{d} in this order. If left 
unspecified, the model will default to \code{c(0, 0, 0, 0)}.}
}}

