% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{linear-class}
\alias{linear-class}
\alias{linear}
\title{An S4 Class to Represent the Linear Model.}
\description{
An S4 Class to Represent the Linear Model.
}
\details{
Defines the \code{linear} class, which states that for a given input \eqn{x}, 
the relationship between \eqn{x} and the dependent variable \eqn{y} is linear:

\deqn{y = a + bx}

This class is solely defined by its parameters \eqn{a} and \eqn{b}, which 
should be provided through the construction of this class.
}
\section{Slots}{

\describe{
\item{\code{parameters}}{Numeric vector containing the values of the parameters of 
the model, namely \eqn{a} and \eqn{b} in this order. If left unspecified, 
the model will default to \code{c(0, 0)}.}
}}

