% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{main_effect-class}
\alias{main_effect-class}
\alias{main_effect}
\title{An S4 Class to Represent the main_effect Effect Model.}
\description{
An S4 Class to Represent the main_effect Effect Model.
}
\details{
Defines the \code{main_effect} class, which states that for two given input
variables \eqn{x} and \eqn{z}, the relationship between these variables and 
the dependent variable \eqn{y} is one with only a main_effect effect:

\deqn{y = a + bx + cz}

This class is solely defined by its parameters \eqn{a}, \eqn{b}, and \eqn{c}, 
which should all be provided through the construction of this class.
}
\section{Slots}{

\describe{
\item{\code{parameters}}{Numeric vector containing the values of the parameters of 
the model, namely \eqn{a}, \eqn{b}, and \eqn{c} in this order. If left 
unspecified, the model will default to \code{c(0, 0, 0)}.}

\item{\code{sd}}{Numeric defining the error around the deterministic part defined by 
the slot \code{parameters}. If left unspecified, will default to \code{1}}
}}

