% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{model-class}
\alias{model-class}
\alias{model}
\title{An S4 Class to Represent all Models.}
\description{
An S4 Class to Represent all Models.
}
\details{
Defines the \code{model} class, to which all other models belong. This class 
is solely defined by its parameters and the error around it.
}
\section{Slots}{

\describe{
\item{\code{parameters}}{Numeric vector containing the values of the parameters of 
the model.}
}}

