% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameter_grid}
\alias{parameter_grid}
\title{Generate parameter grid}
\usage{
parameter_grid(model, n = NA, n_int = 5, n_slope = 5, ...)
}
\arguments{
\item{model}{Object of one of the different model classes (e.g., linear, 
quadratic,...)}

\item{n}{Integer denoting the number of parameter sets to generate. Overwrites
\code{n_int} and \code{n_slope} when specific. Defaults to \code{NA}, meaning 
that \code{n_int} and \code{n_slope} are used by default.}

\item{n_int}{Integer denoting the number of intercepts to draw. Needs to be 
larger than or equal to 2. Defaults to \code{5}.}

\item{n_slope}{Integer denoting the number of slopes to draw. Note that the 
autoregressive parameter is also taken to be a slope. Needs to be 
larger than or equal to 2. Defaults to \code{5}.}

\item{...}{Additional arguments passed on to \code{\link[paramrel]{bounds}}}
}
\value{
Matrix with dimensions N x k, where k is the number of parameters 
of the model and N the total number of combinations possible with these 
parameters
}
\description{
Use the bounds of the parameters of a model and create a grid of parameter 
values. Each parameter value is taken at an equal distance across the interval
of the parameter.
}
