% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramrel-package.R
\docType{package}
\name{paramrel-package}
\alias{paramrel}
\alias{paramrel-package}
\title{paramrel: Parameter Reliability}
\description{
Side project of Kenny and I.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ndpvh/parameter-reliability}
  \item Report bugs at \url{https://github.com/ndpvh/parameter-reliability/issues}
}

}
\keyword{internal}
