% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{prepare}
\alias{prepare}
\title{Prepare True and Estimated Parameters for Analysis}
\usage{
prepare(estimated, simulated)
}
\arguments{
\item{estimated}{Dataframe containing estimated parameters per participant and 
bin, allowing for a decomposition analysis.}

\item{simulated}{Dataframe containing the simulated values for the parameters
per participant.}
}
\value{
Named list returning parameter names (\code{"parameters"}) and 
adjusted data.frames for the estimated and true parameter values 
(\code{"estimated"}, \code{"simulated"}) so that they can be readily compared
to each other.
}
\description{
Prepare True and Estimated Parameters for Analysis
}
