% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{quadratic-class}
\alias{quadratic-class}
\alias{quadratic}
\title{An S4 Class to Represent the Quadratic Model.}
\description{
An S4 Class to Represent the Quadratic Model.
}
\details{
Defines the \code{quadratic} class, which states that for a given input \eqn{x}, 
the relationship between \eqn{x} and the dependent variable \eqn{y} is 
quadratic:

\deqn{y = a + bx + cx^2}

This class is solely defined by its parameters \eqn{a}, \eqn{b}, and \eqn{c},
which should be provided through the construction of this class.
}
\section{Slots}{

\describe{
\item{\code{parameters}}{Numeric vector containing the values of the parameters of 
the model, namely \eqn{a}, \eqn{b}, and \eqn{c} in this order. If left 
unspecified, the model will default to \code{c(0, 0, 0)}.}
}}

