% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{reliability}
\alias{reliability}
\title{Compute the Reliability based on Standard Errors}
\usage{
reliability(data)
}
\arguments{
\item{data}{Named list containing the parameter names under \code{"parameters"},
the true parameter values under \code{"simulated"}, and the estimated 
parameter values under \code{"estimated"}. Typically the result of 
\code{\link[paramrel]{prepare}}}
}
\value{
Dataframe containing the parameter (\code{"parameter"}), the 
coefficient of variation, and the signal-to-noise ratio
}
\description{
Compute the Reliability based on Standard Errors
}
