% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_retest.R
\name{run_testretest_condition}
\alias{run_testretest_condition}
\title{Run Test-Retest Analysis for Single Condition}
\usage{
run_testretest_condition(
  condition_params,
  n_datasets = 100,
  sample_size = 100,
  error_sd = 1
)
}
\arguments{
\item{condition_params}{Parameters for the condition}

\item{n_datasets}{Number of datasets to generate}

\item{sample_size}{Sample size per dataset}

\item{error_sd}{Error standard deviation}
}
\value{
Test-retest reliability results
}
\description{
Run Test-Retest Analysis for Single Condition
}
