% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_retest.R
\name{run_testretest_study}
\alias{run_testretest_study}
\title{Run Complete Test-Retest Study}
\usage{
run_testretest_study(study_size = "pilot", output_dir = "testretest_results")
}
\arguments{
\item{study_size}{Size of study}

\item{output_dir}{Output directory}
}
\value{
Study results
}
\description{
Run Complete Test-Retest Study
}
