% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{signal}
\alias{signal}
\title{Compute the Signal-based Metrics on Standard Errors}
\usage{
signal(data)
}
\arguments{
\item{data}{Named list containing the parameter names under \code{"parameters"},
the true parameter values under \code{"simulated"}, and the estimated 
parameter values under \code{"estimated"}. Typically the result of 
\code{\link[paramrel]{prepare}}}
}
\value{
Dataframe containing the parameter (\code{"parameter"}), the 
coefficient of variation, and the signal-to-noise ratio
}
\description{
Computes the coefficient of variation and the signal-to-noise ratio. The 
coefficient of variation is defined as:

\deqn{CV = \frac{\sigma}{\mu}}

where \eqn{\sigma} is the standard deviation and \eqn{\mu} is the mean of the
estimated parameters. This statistic examines the relationship between the 
observed variation and the average value of the parameters. Note that the 
coefficient of variation is only interpretable if 0 is interpretable on the 
scale. In the case of models, 0 is interpretable on a ratio scale, and so the
coefficient of variation is interpretable as well. However, negative values 
of the parameters are allowed, which is something the coefficient of variation
does not expect, we therefore adjust the formula so that:

\deqn{CV = \frac{\sigma}{|\mu|}}

communicating the relationship between parameter value and standard deviations
in a directionless way.

For random variables, the signal to noise ratio is defined as:

\deqn{SNR = \frac{E[S^2]}{E[N^2]}}

where \eqn{S} represents the signal and \eqn{N} represents the noise.
}
