% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_x}
\alias{simulate_x}
\alias{simulate_x,linear-method}
\alias{simulate_x,quadratic-method}
\alias{simulate_x,cubic-method}
\alias{simulate_x,main_effect-method}
\alias{simulate_x,interaction-method}
\alias{simulate_x,ar1-method}
\alias{simulate_x,arx-method}
\title{Simulate Values for the Independent Variable for a Model}
\usage{
simulate_x(model, ...)

\S4method{simulate_x}{linear}(model, Xfun = function(x) runif(x, -2, 2), N = 100)

\S4method{simulate_x}{quadratic}(model, Xfun = function(x) runif(x, -2, 2), N = 100)

\S4method{simulate_x}{cubic}(model, Xfun = function(x) runif(x, -2, 2), N = 100)

\S4method{simulate_x}{main_effect}(
  model,
  Xfun = list(function(x) runif(x, -2, 2), function(x) runif(x, -2, 2)),
  N = 100
)

\S4method{simulate_x}{interaction}(
  model,
  Xfun = list(function(x) runif(x, -2, 2), function(x) runif(x, -2, 2)),
  N = 100
)

\S4method{simulate_x}{ar1}(model, Xfun = function(x) runif(x, -2, 2), N = 100)

\S4method{simulate_x}{arx}(model, Xfun = function(x) runif(x, -2, 2), N = 100)
}
\arguments{
\item{model}{Object of one of the different model classes (e.g., linear, 
quadratic,...)}

\item{Xfun}{One or more functions with which to simulate values for the independent 
variable(s). Should take in only a single argument, namely \code{N}, 
specifying how many values for \code{X} to simulate. Defaults to a uniform 
distribution bounded between -2 and 2.}

\item{N}{Integer denoting the number of values that should be simulated. 
Defaults to \code{100}.}
}
\value{
Matrix containing the values of the independent variables
}
\description{
Simulate Values for the Independent Variable for a Model
}
