% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_retest.R
\name{test_retest}
\alias{test_retest}
\title{Run Test-Retest Analysis}
\usage{
test_retest(
  sim_model,
  est_model,
  n_participants = 100,
  n_outcomes = 20,
  n_bins = 5,
  parameter_sd = 1,
  icc = 1,
  R2 = NA,
  save_results = TRUE,
  path = file.path("results"),
  filename = "results",
  ...
)
}
\arguments{
\item{sim_model}{Object of the \code{\link[paramrel]{model-class}} containing
parameters around which participant-specific parameters will be generated 
using the \code{\link[paramrel]{generate_parameters}} function. This model 
will be used to simulate data.}

\item{est_model}{Object of the \code{\link[paramrel]{model-class}} that will 
be fitted to the data simulated by \code{sim_model}}

\item{n_participants}{Integer denoting the number of participants to include
in the study}

\item{n_outcomes}{Integer denoting the number of outcomes that exists within
each string. Together with \code{n_bins}, the total number of observations per 
participant then becomes \code{N = n_bins * n_outcomes}. Defaults to 
\code{20}.}

\item{n_bins}{Integer denoting the number of repeated strings of outcomes 
to include in the study. Together with \code{n_outcomes}, the total number of
observations per participant then becomes \code{N = n_bins * n_outcomes}. 
Defaults to \code{5}.}

\item{parameter_sd}{(Positive) numeric denoting the population standard 
deviation of the parameters, therefore indicating how many individual 
differences that exist on each parameter of the simulating model 
\code{sim_model}. Defaults to \code{1}}

\item{icc}{Numeric between 0 and 1 denoting the value of the ICC that you 
want to simulate in the data. If specified, it will change the values of each 
of the parameters in the model across bins, emulating variation across bins  
and therefore the consistency with which they can be recovered. Defaults to 
\code{1}, meaning that parameters don't change across bins.}

\item{R2}{Numeric between 0 and 1 denoting the \eqn{R^2} of the model on the 
within-person level. If specified, it will compute a residual standard 
deviation that allows for the determinstic part of the model to have an 
\eqn{R^2} as specified. Defaults to \code{NA}, triggering the use of the 
model-specified residual standard deviation. Passed on to 
\code{\link[paramrel]{simulate}}}

\item{save_results}{Logical denoting whether to save the results. Defaults to
\code{TRUE}}

\item{path}{Path to the folder where you want to save the different results.
Defaults to a folder \code{"results"} in the current working directory}

\item{filename}{Character denoting the delineating name to be given to the 
files that contain the different results. Defaults to \code{"results"}}

\item{...}{Additional arguments passed on to 
\code{\link[paramrel]{simulate_x}}}
}
\value{
Test-retest reliability results
}
\description{
Given a particular simulation model, estimate a same or other statistical 
model and compute the consistency of the estimated parameters. For this, we 
use the \eqn{ICC(A, 1)}, testing the agreement of the exact values of the 
parameters within participants.
}
\details{
This function assesses the test-retest consistency of a set of estimated 
parameters in a particular setting. Specifically, we assume that a same 
participant has a same set of parameters that underlie their responses. With
this assumption in place, we then simulate a dataset where a participant 
encounters the same string of values for the independent variable(s) 
multiple times throughout the study. By estimating a statistical model for 
each participant and each repeated string of outcomes separately, we can 
then assess the test-retest consistency of these estimated parameters, which
is the goal of this study

As the operationalization of consistency, we compute the \eqn{ICC(A, 1)} of 
the parameters across participants, in our context being defined as:

\deqn{ICC(A, 1) = \frac{\sigma_\text{participants}}^2}{\sigma_\text{total}^2}

where \eqn{\sigma^2} denotes the variance. This \eqn{ICC} puts the systematic 
variance -- that is, the interindividual differences in the values of the 
parameters -- against the total observed variance in the parameters, checking 
whether individual differences can be reliably picked up on.
}
