% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{ar1-class}
\alias{ar1-class}
\alias{ar1}
\title{An S4 Class to Represent the ar1 Model.}
\description{
An S4 Class to Represent the ar1 Model.
}
\details{
Defines the \code{ar1} class, which states that for a given dependent
variable \eqn{y}, the previous values of this variable in part determine 
current values of the variable:

\deqn{y_{t} = a + by_{t - 1}}

This class is solely defined by its parameters \eqn{a} and \eqn{b}, which 
should all be provided through the construction of this class.
}
\section{Slots}{

\describe{
\item{\code{parameters}}{Numeric vector containing the values of the parameters of 
the model, namely \eqn{a} and \eqn{b} in this order. If left unspecified, the 
model will default to \code{c(0, 0)}.}

\item{\code{sd}}{Numeric defining the error around the deterministic part defined by 
the slot \code{parameters}. If left unspecified, will default to \code{1}}
}}

