% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{bounds}
\alias{bounds}
\alias{bounds,linear-method}
\alias{bounds,quadratic-method}
\alias{bounds,cubic-method}
\alias{bounds,main-method}
\alias{bounds,interaction-method}
\alias{bounds,ar1-method}
\alias{bounds,arx-method}
\title{Get the Bounds of the Model}
\usage{
bounds(model, ...)

\S4method{bounds}{linear}(model, intercept = c(-10, 10), slope = c(-5, 5))

\S4method{bounds}{quadratic}(model, intercept = c(-10, 10), slope = c(-5, 5))

\S4method{bounds}{cubic}(model, intercept = c(-10, 10), slope = c(-5, 5))

\S4method{bounds}{main}(model, intercept = c(-10, 10), slope = c(-5, 5))

\S4method{bounds}{interaction}(model, intercept = c(-10, 10), slope = c(-5, 5))

\S4method{bounds}{ar1}(model, intercept = c(-10, 10), slope = c(-5, 5))

\S4method{bounds}{arx}(model, intercept = c(-10, 10), slope = c(-5, 5))
}
\arguments{
\item{model}{Object of one of the different model classes (e.g., linear, 
quadratic,...)}
}
\value{
Matrix where the left/right column specifies the lower/upper bounds 
of the parameters in the order specified in the description of the models
}
\description{
Get the Bounds of the Model
}
