% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{initialize,linear-method}
\alias{initialize,linear-method}
\title{Constructor for the \code{\link[paramrel]{linear-class}}}
\usage{
\S4method{initialize}{linear}(.Object, parameters = c(0, 0), sd = 1)
}
\arguments{
\item{.Object}{Model to initialize}

\item{parameters}{Numeric vector containing the values of the parameters of 
the model, namely \eqn{a} and \eqn{b} in this order. If left unspecified, 
the parameters will default to \code{c(0, 0)}.}

\item{sd}{Numeric defining the error around the deterministic part defined by 
the argument \code{parameters}. If left unspecified, will default to \code{1}}
}
\description{
Constructor for the \code{\link[paramrel]{linear-class}}
}
