% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\alias{simulate,linear-method}
\alias{simulate,quadratic-method}
\alias{simulate,main-method}
\alias{simulate,interaction-method}
\alias{simulate,ar1-method}
\alias{simulate,arx-method}
\title{Simulate Data from a Model}
\usage{
simulate(model, ...)

\S4method{simulate}{linear}(model, X = NULL, Xfun = function(x) runif(x, -2, 2), N = 100)

\S4method{simulate}{quadratic}(model, X = NULL, Xfun = function(x) runif(x, -2, 2), N = 100)

\S4method{simulate}{main}(model, X = NULL, Xfun = function(x) runif(x, -2, 2), N = 100)

\S4method{simulate}{interaction}(model, X = NULL, Xfun = function(x) runif(x, -2, 2), N = 100)

\S4method{simulate}{ar1}(model, X = NULL, Xfun = function(x) runif(x, -2, 2), N = 100)

\S4method{simulate}{arx}(model, X = NULL, Xfun = function(x) runif(x, -2, 2), N = 100)
}
\arguments{
\item{model}{Object of one of the different model classes (e.g., linear, 
quadratic,...)}

\item{X}{Numeric vector or matrix containing the values of the independent 
variables. For \code{\link[param-rel]{linear-class}}, 
\code{\link[param-rel]{quadratic-class}}, \code{\link[param-rel]{arx-class}},
a numeric vector suffices, as only one independent variable is required. For 
\code{\link[param-rel]{main-class}} and 
\code{\link[param-rel]{interaction-class}}, a matrix is needed instead, where 
the first column is taken to be \eqn{x} (paired with parameter \eqn{b}) and 
the second column is taken to be \eqn{z} (paired with parameter \eqn{c}). For
the \code{\link[param-rel]{ar1-class}}, this argument is ignored.
Defaults to \code{NULL}, forcing the user to specify its value for all models
except for the \code{\link[param-rel]{ar1-class}}, or alternatively
to specify a function with which to generate values of \code{X} through the
\code{Xfun} argument.}

\item{Xfun}{Function with which to simulate values for \code{X}. Should take 
in only a single argument, namely \code{N}, specifying how many values for
\code{X} to simulate. Defaults to \code{NULL}, forcing the user to either 
specify \code{X} or \code{Xfun}. Note that the output of \code{Xfun} is used 
as an alternative to \code{X}, meaning that it should conform to distinction 
between numeric vector and numeric matrix specified for the \code{X} argument.}

\item{N}{Integer denoting the number of values that should be simulated. 
Ignored when \code{X} is defined. Defaults to \code{100}.}
}
\value{
Dataframe containing the values of the variables (\eqn{y}, \eqn{x}, 
and if applicable \eqn{z}, named as such) and time (\code{time})
}
\description{
Method for the different models that allows the simulation of data for these
models.
}
